+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file. Be aware th
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Items first seen in the bank no longer end up in the wrong category after being withdrawn
+
 ## [2.0.1] 2026-09-13
 ### Fixed
 - BoE, WuE and BoA items no longer go missing from their categories after logging in